1. added README 2. refactored registerUser method in RegistrationController 3. transferred package annotations and files EmailValidator , PasswordMatchesValidator to package lib.
