+++ README.md
@@ -1,3 +1,54 @@
-# jv-spring-hibernate-template
+# Library
 
-In order to run this project please replace the data with proper values in db.properties
+# Table of Contents
+
+* [Project description](#description)
+* [Stack of technologies](#technologies)
+* [How to start?](#start)
+* [Authors](#authors)
+
+# <a name="description"></a>Project description
+This project is a draft for your own online library.
+
+It allows user to:
+- register as a new user
+- login using  authentication and authorization
+- find a book by title (or id)
+- rent book
+- return book
+- see all rented books
+- see all books
+
+# <a name="technologies"></a>Stack of technologies
+* Java 12
+* Maven 4.0.0
+* MavenCheckstyle 3.1.0
+* Tomcat 8.5.45
+* MySQL 8.0.17
+* Spring context 4.3.7
+* Hibernate 5.2.9.Final
+* javax.annotation-api 1.3.2
+* jaxb-api 2.3.1
+
+# <a name="start"></a>How to start?
+
+Open this project in your IDE.
+
+Make project as maven project.
+
+Install and configure Tomcat.
+
+Install and configure database management system.
+
+In src.main.resources.db.properties class you have to change user and password for you DB.
+
+You can run this project.
+
+By default you have user with role USER and ADMIN (login = admin, password = 1). 
+
+# <a name="authors"></a>Authors
+Author: Tatiana Pyrch
+
+LinkedIn: [My LinkedIn Profile](https://www.linkedin.com/in/tatiana-pyrch-97235b162)
+
+GitHub: [My Projects](https://github.com/TatianaPirch)
